pyproject.toml
[project]
name = "tyre-combiner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "debugpy>=1.8.19",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
quote-style = "double"

[tool.basedpyright]
venvPath = "."
venv = ".venv"
useLibraryCodeForTypes = true
typeCheckingMode = "basic"  # Временно отключаем строгую проверку
analysis.diagnosticMode = "openFiles"
python.analysis.autoImportCompletions = true



[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "basedpyright>=1.20.0",
]
